% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check_functions.R
\name{arg_check_alternative}
\alias{arg_check_alternative}
\title{Argument check alternative}
\usage{
arg_check_alternative(alternative)
}
\arguments{
\item{alternative}{One of "lower", "greater", "two.sided"}
}
\description{
Argument check alternative
}
