% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.clusters.R
\name{clusters.spscan}
\alias{clusters.spscan}
\title{Extract clusters}
\usage{
clusters.spscan(x, idx = seq_along(x$clusters), ...)
}
\arguments{
\item{x}{An object of class \code{spscan} from the \code{\link{spscan.test}}}
}
\value{
A list. Each element of the list is a vector with the indices of event
locations in the associated cluster.
}
\description{
Extract clusters
}
\examples{
data(grave)
# apply scan method
out = spscan.test(grave, nsim = 99)
# print scan object
clusters(out)
}
