% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdest-print.R
\name{print.kdenv}
\alias{print.kdenv}
\title{Print a \code{kdenv} object}
\usage{
\method{print}{kdenv}(x, ..., extra = FALSE)
}
\arguments{
\item{x}{An object produced by the
\code{\link[smacpod]{kdest}} function.}

\item{...}{Not currently implemented.}

\item{extra}{A logical value indicating whether extra
information related to the internal
\code{\link[spatstat.core]{fv}} object should be printed.
The default is \code{FALSE}.}
}
\value{
Information about the \code{kdest}
}
\description{
Print an \code{kdenv} object produced by
\code{\link[smacpod]{kdest}}.
}
\author{
Joshua French
}
