% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdplus.test.R
\name{print.kdplus_test}
\alias{print.kdplus_test}
\title{Print a \code{kdplus_test} object}
\usage{
\method{print}{kdplus_test}(x, ...)
}
\arguments{
\item{x}{An object produced by the \code{\link[smacpod]{kdplus.test}} function.}

\item{...}{Not currently implemented.}
}
\value{
Information about the test
}
\description{
Print a \code{kdplus_test} object produced by 
\code{\link[smacpod]{kdplus.test}}.
}
\author{
Joshua French
}
