% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrr.test.R
\name{print.logrr_test}
\alias{print.logrr_test}
\title{Print a \code{logrr_test} object}
\usage{
\method{print}{logrr_test}(x, ...)
}
\arguments{
\item{x}{An object produced by the \code{\link[smacpod]{logrr.test}} function.}

\item{...}{Not currently implemented.}
}
\value{
Information about the test
}
\description{
Print an \code{logrr_test} object produced by
\code{\link[smacpod]{logrr.test}}.
}
\author{
Joshua French
}
