% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spscan.R
\name{print.spscan}
\alias{print.spscan}
\title{Plots object from \code{\link{spscan.test}}.}
\usage{
\method{print}{spscan}(x, ..., extra = FALSE)
}
\arguments{
\item{x}{An object of class \code{spscan}.}

\item{...}{Additional arguments affecting the summary produced.}

\item{extra}{A logical value. Default is \code{FALSE}.
\code{TRUE} indicates that extra information should be
printed.}
}
\description{
Plots object of class \code{spscan} from
\code{\link{spscan.test}}.
}
\details{
If \code{border}, \code{ccol}, \code{clty}, or 
\code{clwd} are specified, then the length of these
vectors must match \code{nrow(x$coords)}.
}
\examples{
data(grave)
out = spscan.test(grave, case = 2, alpha = 0.1)
out
}
