% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnn.test.R
\name{qnn.test}
\alias{qnn.test}
\title{q Nearest Neighbors Test}
\usage{
qnn.test(x, q = 5, case = 2, nsim = 499, longlat = FALSE)
}
\arguments{
\item{x}{A \code{\link[spatstat.geom]{ppp}} object with marks for the case
and control groups.}

\item{q}{A vector of positive integers indicating the
values of \code{q} for which to do the q nearest
neighbors test.}

\item{case}{The name of the desired "case" group in
\code{levels(x$marks)}. Alternatively, the position of
the name of the "case" group in \code{levels(x$marks)}.
Since we don't know the group names, the default is 2,
the second position of \code{levels(x$marks)}.
\code{x$marks} is assumed to be a factor.  Automatic
conversion is attempted if it is not.}

\item{nsim}{The number of simulations from which to
compute p-value.}

\item{longlat}{A logical value indicating whether
Euclidean distance (\code{FALSE}) or Great Circle
(WGS84 ellipsoid, \code{FALSE}) should be used. Default
is \code{FALSE}, i.e., Euclidean distance.}
}
\value{
Returns a list with the following components:
  \item{qsum}{A dataframe with the number of neighbors
  (q), test statistic (Tq), and p-value for each test.}
  \item{consum}{A dataframe with the contrasts
  (contrast), test statistic (Tcon), and p-value
  (pvalue) for the test of contrasts.}
}
\description{
\code{qnn.test} calculates statistics related to the q
nearest neighbors method of comparing case and control
point patterns under the random labeling hypothesis.
}
\examples{
data(grave)
qnn.test(grave, case = "affected", q = c(3, 5, 7, 9, 11, 13, 15))
}
\references{
Waller, L.A., and Gotway, C.A. (2005).
  Applied Spatial Statistics for Public Health Data.
  Hoboken, NJ: Wiley.

  Cuzick, J., and Edwards, R. (1990). Spatial clustering
  for inhomogeneous populations. Journal of the Royal
  Statistical Society. Series B (Methodological), 73-104.

  Alt, K.W., and Vach, W. (1991). The reconstruction of
  "genetic kinship" in prehistoric burial
  complexes-problems and statistics. Classification, Data
  Analysis, and Knowledge Organization, 299-310.
}
\author{
Joshua French
}
