% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smacpod-package.R
\docType{package}
\name{smacpod}
\alias{smacpod-package}
\alias{smacpod}
\title{smacpod}
\description{
*S*tatistical *M*thods for the *A*nalysis of *C*ase-control *Po*int *D*ata
}
\details{
Statistical methods for analyzing case-control point
data. Methods include the ratio of kernel densities, the
difference in K Functions, the spatial scan statistic,
and q nearest neighbors of cases.

Run
}
\author{
\strong{Maintainer}: Joshua French \email{joshua.french@ucdenver.edu} (\href{https://orcid.org/0000-0002-9708-3353}{ORCID})

}
