% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spscan.test.R
\name{spscan.sim}
\alias{spscan.sim}
\title{Compute spatial scan statistics for simulated data}
\usage{
spscan.sim(nsim, N, N1, Nin, Nout, const, cl)
}
\arguments{
\item{nsim}{The number of simulations from which to compute the p-value.  A
non-negative integer.  Default is 499.}

\item{N}{Number of event locations}

\item{N1}{Number of case events}

\item{Nin}{Number of event locations in each window}

\item{Nout}{Number of event locations outside each window}

\item{const}{Constant for computing test statistic}

\item{cl}{
A cluster object created by \code{\link{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations
(see Details on performance).
It can also be \code{"future"} to use a future backend (see Details),
\code{NULL} (default) refers to sequential evaluation.
}
}
\value{
A vector of maximum test statistics
}
\description{
Compute spatial scan statistics for simulated data
}
\keyword{internal}
