% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spscan.test.R
\name{spscan.stat}
\alias{spscan.stat}
\title{Compute spatial scan statistic}
\usage{
spscan.stat(N, N1, Nin, Nout, N1in, const)
}
\arguments{
\item{N}{Number of event locations}

\item{N1}{Number of case events}

\item{Nin}{Number of event locations in each window}

\item{Nout}{Number of event locations outside each window}

\item{N1in}{Number of case events in each window}

\item{const}{Constant for computing test statistic}
}
\value{
A numeric vector of test statistics
}
\description{
Compute spatial scan statistic
}
\keyword{internal}
