% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdest-summary.R
\name{summary.kdenv}
\alias{summary.kdenv}
\title{Summarize a \code{kdenv} object}
\usage{
\method{summary}{kdenv}(object, ..., extra = FALSE)
}
\arguments{
\item{object}{An object produced by the \code{\link[smacpod]{kdest}}
function.}

\item{...}{Not currently implemented.}
}
\value{
A list that contains the sequences of indices for which the estimated
  difference in KD functions is above the envelopes, below the envelopes, and
  the vector of distances.
}
\description{
Summarize the sequences of distances for which the difference in estimated K
functions, \code{KD(r) = K_case(r) - K_control(r)}, falls outside the
non-rejection envelopes.
}
\author{
Joshua French
}
