% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grave.R
\docType{data}
\name{grave}
\alias{grave}
\title{Medieval Grave Site Data}
\format{ppp (planar point process)}
\source{
Waller, L.A. and Gotway, C.A. (2005).  Applied Spatial Statistics for Public Health Data.  Hoboken, NJ: Wiley.
}
\usage{
grave
}
\description{
This data set contains 143 observations of medieval grave site data stored as a \code{ppp} object from the \code{spatstat} package.  The data are marked as being "affected" by a tooth deformity or "unaffected" by a tooth deformity.
}
\author{
Joshua French
}

