% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grave.R
\docType{data}
\name{grave}
\alias{grave}
\title{Medieval Grave Site Data}
\format{
\code{ppp} (planar point process) class object
  from the \code{spatstat} package.
}
\source{
Waller, L.A. and Gotway, C.A. (2005).  Applied
  Spatial Statistics for Public Health Data.  Hoboken,
  NJ: Wiley.
}
\usage{
data(grave)
}
\description{
This data set contains 143 observations of
  medieval grave site data stored as a \code{ppp} class
  object from the \code{spatstat} package.  The data are
  marked as being "affected" by a tooth deformity or
  "unaffected" by a tooth deformity.
}
\seealso{
\code{\link[spatstat.geom]{ppp}}
}
\author{
Joshua French
}
