% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kdenv.R
\name{plot.kdenv}
\alias{plot.kdenv}
\title{Plots object from kdest of class \code{kdenv}.}
\usage{
\method{plot}{kdenv}(x, ..., shadecol1 = "grey", shadecol2 = "lightblue",
  main = "", legend = FALSE)
}
\arguments{
\item{x}{An object of class kdenv to be plotted.}

\item{...}{Additional graphical parameters passed to \code{plot.fv} function.}

\item{shadecol1}{Shade color for max/min envelopes.}

\item{shadecol2}{Shade color for confidence envelopes.}

\item{main}{A main title for the plot.  Default is blank.}

\item{legend}{Logical for whether a legend should automatically be displayed.  Default if \code{FALSE}.}
}
\description{
Plots object from kdest of class \code{kdenv}.
}
\examples{
data(grave)
kd1 = kdest(grave, nsim = 19, level = 0.9)
plot(kd1)
}
\seealso{
\code{\link[spatstat]{plot.fv}}
}
